docs(strategies): Refine docs in code and md
